Add FileOperation

+++ src/main/java/eu/mytthew/Plugin.java
@@ -4,6 +4,7 @@ import org.json.JSONObject;
 
 import java.io.IOException;
 import java.security.GeneralSecurityException;
+import java.time.LocalDate;
 import java.util.concurrent.ScheduledThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
 
@@ -17,17 +18,23 @@ public class Plugin {
 		exec.schedule(() -> {
 			try {
 				int description = 1;
+				String todayDate = LocalDate.now().toString();
 				JSONObject jsonObject = new JSONObject();
 				jsonObject.put("description", description);
 				if (fileOperation.fileExist("last")) {
 					jsonObject = fileOperation.openFile("last");
-					description = jsonObject.getInt("description");
-					description = description + 1;
-					fileOperation.deleteFile("last");
+					if (!jsonObject.getString("date").equals(todayDate)) {
+						description = jsonObject.getInt("description");
+						fileOperation.deleteFile("last");
+						eventManager.addEvent(description + 1);
+						jsonObject.put("date", todayDate);
+						fileOperation.createFile("last", jsonObject);
+					}
+				} else {
+					eventManager.addEvent(description);
+					jsonObject.put("date", todayDate);
+					fileOperation.createFile("last", jsonObject);
 				}
-				eventManager.addEvent(description);
-				fileOperation.createFile("last", jsonObject);
-
 			} catch (IOException | GeneralSecurityException e) {
 				e.printStackTrace();
 			}
